Move sessions to a component.
