+++ app/api/v1/api.py
@@ -1,8 +1,16 @@
 from fastapi import APIRouter
 
 from app.api.v1.endpoints import geocoding
-from app.api.v1.endpoints import sessions
-from app.components import locations, oauth, zones, auth, organizations, users, guests
+from app.components import (
+    locations,
+    oauth,
+    zones,
+    auth,
+    organizations,
+    users,
+    guests,
+    sessions,
+)
 
 api_router = APIRouter()
 # TODO: Consider moving prefixes and tags to router declarations.
@@ -14,7 +22,7 @@ api_router.include_router(users.routes.router, prefix="/users", tags=["users"])
 api_router.include_router(
     organizations.routes.router, prefix="/organizations", tags=["organizations"]
 )
-api_router.include_router(sessions.router, prefix="/sessions", tags=["sessions"])
+api_router.include_router(sessions.routes.router, prefix="/sessions", tags=["sessions"])
 api_router.include_router(oauth.routes.router, prefix="/oauth", tags=["oauth"])
 api_router.include_router(zones.routes.router, prefix="/zones", tags=["zones"])
 api_router.include_router(guests.routes.router, prefix="/guest", tags=["guest-user"])
