+++ app/components/sessions/crud.py
@@ -2,7 +2,7 @@ from typing import Optional, List
 
 from sqlalchemy.orm import Session
 
-from app.models.sessionhistory import SessionHistory
+from app.components.sessions import models
 
 
 def create(
@@ -11,10 +11,10 @@ def create(
     user_id: int,
     access_token: str,
     user_agent: str = None,
-    user_ip: str = None
+    user_ip: str = None,
 ):
 
-    db_obj = SessionHistory(
+    db_obj = models.SessionHistory(
         user_id=user_id,
         access_token=access_token,
         user_agent=user_agent,
@@ -28,29 +28,41 @@ def create(
 
 def get_by_access_token(
     db: Session, *, user_id: int, access_token: str
-) -> Optional[SessionHistory]:
+) -> Optional[models.SessionHistory]:
     return (
-        db.query(SessionHistory)
+        db.query(models.SessionHistory)
         .filter(
-            SessionHistory.access_token == access_token,
-            SessionHistory.user_id == user_id,
+            models.SessionHistory.access_token == access_token,
+            models.SessionHistory.user_id == user_id,
         )
         .first()
     )
 
 
-def get_user_active_sessions(db: Session, *, user_id: int) -> List[SessionHistory]:
+def get_user_active_sessions(
+    db: Session,
+    *,
+    user_id: int,
+) -> List[models.SessionHistory]:
     return (
-        db.query(SessionHistory)
-        .filter(SessionHistory.user_id == user_id, SessionHistory.is_active == True)
+        db.query(models.SessionHistory)
+        .filter(
+            models.SessionHistory.user_id == user_id,
+            models.SessionHistory.is_active == True,
+        )
         .all()
     )
 
 
-def revoke_by_id(db: Session, *, user_id: int, session_id: int) -> SessionHistory:
+def revoke_by_id(
+    db: Session, *, user_id: int, session_id: int
+) -> models.SessionHistory:
     db_obj = (
-        db.query(SessionHistory)
-        .filter(SessionHistory.user_id == user_id, SessionHistory.id == session_id)
+        db.query(models.SessionHistory)
+        .filter(
+            models.SessionHistory.user_id == user_id,
+            models.SessionHistory.id == session_id,
+        )
         .first()
     )
     if not db_obj:
