+++ app/db/base.py
@@ -5,10 +5,10 @@ from app.components.locations.models import Location
 from app.components.oauth.models import OauthScope, OauthRole, association_table
 from app.components.organizations.models import Organization
 from app.components.reports.models import Report, ReportOption
+from app.components.sessions.models import SessionHistory
 from app.components.users.models import User
 from app.components.zones.models import Zone
 from app.db.base_class import Base
-from app.models.sessionhistory import SessionHistory
 
 __all__ = [
     "Base",
