+++ app/schemas/__init__.py
@@ -1,5 +1,3 @@
 from .enums import *
 from .geocoding import *
 from .roles import UserRole
-from .session import UserSession
-from .session import UserSession
